LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/page-load/meta-referrer-no-referrer-img-src.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: expected the same number of headers between speculative and non-speculative expected 2 but got 3


FAIL Speculative parsing, page load: meta-referrer-no-referrer-img-src Unhandled rejection: assert_equals: expected the same number of headers between speculative and non-speculative expected 2 but got 3
